Refactored code into timer/src/utils.rs.
